move data model tests
